requirements.txt
asgiref==3.5.2
boto3==1.20.46
Django==4.0.6
django-allauth==0.51.0
django-countries==7.2.1
django-debug-toolbar==3.2.4
django-hitcount==1.3.5
django-storages==1.12.3
gunicorn==20.1.0
Pillow==9.2.0

